Changed Index (default) to be Landing page. Renamed existing Index to Home
