+++ food-for-thought/app/_layout.tsx
@@ -5,11 +5,11 @@ export default function RootLayout() {
   return (
       <GestureHandlerRootView style={{flex: 1}}>
           <Stack>
-            <Stack.Screen name="index" options={{headerShown:false,fullScreenGestureEnabled:true,gestureDirection:'horizontal'}}/>
+            <Stack.Screen name="index" options={{headerShown:false}}/>
             <Stack.Screen name="map" options={{headerShown:false,fullScreenGestureEnabled:true,gestureDirection:'horizontal'}}/>
             <Stack.Screen name="login" options={{headerShown:false,fullScreenGestureEnabled:true,gestureDirection:'horizontal'}}/>
             <Stack.Screen name="register" options={{headerShown:false,fullScreenGestureEnabled:true,gestureDirection:'horizontal'}}/>
-            <Stack.Screen name="landing" options={{headerShown:false}}/>
+            <Stack.Screen name="home" options={{headerShown:false,fullScreenGestureEnabled:true,gestureDirection:'horizontal'}}/>
           </Stack>
       </GestureHandlerRootView>
   );
